os: Implement newtype for HRESULT with must_use and hexadecimal format
Unify the use of `com_rs::HResult` into a newtyped `HRESULT` that
allows us to place a `#[must_use]` annotation and in turn prevent
missing errors from the API, which could lead to worse Undefined
Behaviour.

In addition we can now use hexadecimal formatting for the `Display` and
`Debug` traits making them much easier to read in `unwrap()` and similar
scenarios.