Convert return_hr to expression
Make return_hr more versatile by providing Ok or Err as expession
instead of forcing a return out of the current function. This follows
Rusts convention to provide a block result as an unterminated expression
on the last line, and allows it to be used in an expression context
including with the question mark operator.