Linux: DXC BSTR now resembles its Windows counterpart with explicit size
We [recently] fixed memory leaks when calling into the intellisense part
of DXC and made the decision shortly after merging to [implement
null-invariant `BSTR`s properly] in DXC's shimmed type for non-Windows,
which was promptly accepted.

Currently our implementation of `SysFreeString` together with that PR
segfaults because it frees the pointer at offset `+4` of the allocation.
That is corrected together with `SysStringLen` now reading the size of
the string from this prefix instead of looking for a (wrong!) null
termination character.  At least our Windows and Linux implementation
`utils.rs` is the same again.

[recently]: https://github.com/Traverse-Research/hassle-rs/pull/10
[implement null-invariant `BSTR`s properly]: https://github.com/microsoft/DirectXShaderCompiler/pull/3250