Free leaking DxcIncludeHandlerWrapper
Valgrind shows that we are leaking the DxcIncludeHandlerWrapper
allocation:

    ==96758== 485 (80 direct, 405 indirect) bytes in 1 blocks are definitely lost in loss record 13 of 17
    ==96758==    at 0x483A77F: malloc (vg_replace_malloc.c:307)
    ==96758==    by 0x1A1E4B: alloc::alloc::alloc (alloc.rs:74)
    ==96758==    by 0x1A1F09: alloc::alloc::Global::alloc_impl (alloc.rs:153)
    ==96758==    by 0x1A2069: <alloc::alloc::Global as core::alloc::AllocRef>::alloc (alloc.rs:203)
    ==96758==    by 0x1A1D9A: alloc::alloc::exchange_malloc (alloc.rs:281)
    ==96758==    by 0x199A1A: new<hassle_rs::wrapper::DxcIncludeHandlerWrapper> (boxed.rs:175)
    ==96758==    by 0x199A1A: hassle_rs::wrapper::DxcCompiler::prep_include_handler (wrapper.rs:248)
    ==96758==    by 0x199D6F: hassle_rs::wrapper::DxcCompiler::compile (wrapper.rs:278)
    ==96758==    by 0x195C70: hassle_rs::utils::compile_hlsl (utils.rs:115)
    ==96758==    by 0x127BDF: include::main (autogen_ops.rs:0)
    ==96758==    by 0x1275CA: core::ops::function::FnOnce::call_once (autogen_context.rs:1683)
    ==96758==    by 0x127ADD: std::sys_common::backtrace::__rust_begin_short_backtrace (backtrace.rs:137)
    ==96758==    by 0x127530: std::rt::lang_start::{{closure}} (rt.rs:66)

This happens because of Box::into_raw, which moves the pointer out of
the box. Taking this pointer out of a referenced Box makes it live on
till the end of the function where it is appropriately dropped.

Properly freeing DxcIncludeHandlerWrapper exposes a second issue: The
blobs are now freed but have already been freed within DXC becaues of a
missing refcount: we are supposed to increment it before writing the
pointer to it to `include_source` [1].
Now that the blobs have a proper refcount and live on until DXC cleans
them up when not needed, they do not need to be kept alive inside
DxcIncludeHandlerWrapper anymore.

Finally, clean up the unnecessary clone of `source`: This String is
already retrieved by value and can be moved straight into the Rc.

[1]: https://github.com/microsoft/DirectXShaderCompiler/blob/a8d9780046cb64a1cea842fa6fc28a250e3e2c09/tools/clang/unittests/dxc_batch/dxc_batch.cpp#L553-L554