wrapper: Pin blob source insted of unnecessarily RC'ing it (#14)
The refcounter here makes little sense as the string is only ever
referenced by DxcIncludeHandlerWrapper.  The count is never artificially
inremented to represent implicit referencing from a native DxcBlob.
Overall lifetime of the string is not altered: as soon as
DxcIncludeHandlerWrapper goes out of scope the underlying string
disappears as well.

In addition Clippy doesn't like this extra, unnecssary level of
indirection either, when not on the inner (`str`) type:

    warning: usage of `Rc<T>` when T is a buffer type
       --> src/wrapper.rs:139:17
        |
    139 |     pinned: Vec<Rc<String>>,
        |                 ^^^^^^^^^^ help: try: `Rc<str>`
        |
        = note: `#[warn(clippy::rc_buffer)]` on by default
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#rc_buffer

Furthermore, use Pin<> (following the name of the variable) to hold the
string at the same location in memory.  In particular this Pin instance
only gives us access to the Deref of `String` which is `&str` and does
not allow destructive operations like resizing or reassignment that
could potentially reallocate heap memory (hence moving the pointer).