Revert "[WIN/LINUX] Implement vtable dummies (virtual dtors) through extra trait"
This reverts commit c37cb037be072cac43f2706b89f1388b9deee19a. [This
commit] is squashed in the merge of PR #5, and does not exist in-tree.

https://github.com/microsoft/DirectXShaderCompiler/pull/3793 has finally
been merged on October 13th 2022 and now allows us to drop the vtable
dummies for virtual destructors again, as those are not part of the ABI
and shouldn't have been in the vtable in the first place.

[This commit]: https://github.com/Traverse-Research/hassle-rs/pull/5/commits/c37cb037be072cac43f2706b89f1388b9deee19a