utils: Provide `OperationOutput` wrapper to parse `DxcOperationResult`
Thus far `hassle-rs` has made it possible - but not very convenient -
to extract warnings/errors from a `DxcOperationResult` together with
an optional successful result-blob.  This was always eaten up by our
high-level `utils` wrappers however, giving you either errors on failure
or only the result blob on success, but never the warnings.  This
was always only available when manually calling the high level `ffi`
`wrapper`, with the same inconvenience that `utils` has.

There's barely any upstream documentation, but through experimentation
we see that both the output and error blobs are always returned, and
`get_status()` is the definitive answer on whether the compilation was
successful.  In either case, warnings/errors are available in the error
buffer if its of non-zero length.

Thus we improve `HassleError` to contain the error string next to the
(typically `E_FAIL`) `HRESULT`, leaving the new `OperationOutput` to
contain the blob and all optional warnings/messages.

This parsing is still optional when using the `wrapper` API, but easy to
call by downstream crates if they wish to simplify their handling into
a `match` statement.  The `utils` wrappers return this type directly,
incurring a little bit of "allocation overhead" for copying the
underlying data into `Vec`s.

Note that this behaviour relies on previous patches fixing our UB
`as_slice()` implementations to no longer construct (empty!) slices from
NULL pointers.