examples/fake-signing: Make runnable on Linux
The main reason for fake signing is making it run on Linux - and not
depend on `dxil.dll` on Windows - so it's great if this example runs on
Linux as well to see it all working even if that requires a few nasty
cfg blocks.  Note that this is done with `cfg!()` and regular `if`
blocks instead of attributes, such that this code lints and compiles on
both Windows and Linux even if the `validate_dxil()` codepath is never
entered on the latter (and that way includes, `let mut all_matches` etc
don't need to be `#[cfg(windows)]`'ed out either).