Consistently return `HassleError` and remove unneeded check macros
Our public API becomes a lot cleaner when returning `HassleError`
everywhere, helped by the fact that `HRESULT`s from the low-level COM
API is a newtype with support for functions to convert to `Result<>`
easily.  These can be short-circuit-returned in a more Rust'y way
through the questionmark operator instead of wrapping everything in
large macro calls.  This change is based on a similar approach in the
Ash crate: https://github.com/MaikKlein/ash/pull/339/files