Use `.cast()` to not accidentally const_cast, fix Android char type
Clean up casts to use `.cast()` to change the type while making it
impossible to accidentally turn a `*const` into a `*mut`.

Also makes sure that the correct const-ness types, ie lpCstr and friends
are used when the pointed-to data needs not to be modified: `*mut`
coerces to `*const` (but for obvious reasons not the other way around).

Finally, fix the `CHAR` type to use `std::os::raw::c_char` which is
defined to `u8` on Arm architectures and i8 on our x64 machines.