ci: Drop nightly build/lint tests (#34)
There is not much point "preparing hassle-rs for the future" if the
lints are false positives and prevent us from submitting or having
actually good code.  Overnight two false positives got introduced:
https://github.com/Traverse-Research/hassle-rs/runs/5006913943?check_suite_focus=true.
One needs the borrow to satisfy a peculiar `From`/`Into` implementation,
and the `&mut Vec<>` is pushed to which makes it impossible to turn this
into a mutable borrow of a slice.