Migrate to Microsoft's com-rs (#35)
The `com-rs` crate is deprecated, but Microsoft published their vetted
and maintained replacement as `com`.  This crate has much better helpers
and doesn't pull in the `libuuid` library that isn't available on
Android nor Mac.

Unfortunately there is still no verdict how this crate relates to the
windows-rs iniative, where the entire FFI API is autogenerated from
metadata, nor has it been possible to bring in Linux support thus far.
As such we'll stick with Microsoft's `com` for now which is already a
step in the right direction.

Note that we'll have to figure out a cleaner solution to get a
stack-allocated, temporary class allocation for the include handler: one
that is only allowed to live as long as the surrounding function,
instead of endlessly on the heap.