wrapper: Reimplement `DxcBlob` "casting" via `transmute()`s
Because `DxcBlobEncoding` inherits directly from `DxcBlob`, we're
allowed to move ownership from one to the other.  Especially without
creating a clone via `query_interface()` only to drop the source
right after. (where cloning and dropping refer to incrementing and
decrementing the object refcount, respectively)

Note that this relies on our wrapper objects to be `repr(transparent)`,
not have any other fields, and not have custom `Drop` containers,
otherwise we'd have to construct a new wrapper (that `Deref` cannot
return as borrow!) to store a `transmute()`d `ffi::IDxcBlobEncoding` in.