wrapper: Load libraries globally instead of relative to `$PWD` (#74)
There's no real reason to force `./` in default paths when the library
is more and more likely to be installed on the system via `PATH` (on
Windows it's in the Vulkan SDK, on Linux `directx-shader-compiler` is
a common package).

Furthermore, stop appending this path to the optional `lib_path`
argument, if that `lib_path` doesn't **exist** as a file.  In the case
that a user specifies a (relative) path that doesn't exist, this throws
an obscure error that `libdxcompiler.so/./libdxcompiler.so` doesn't
exist because we appended a precooked filename a second time.