Fix missing error log in `HassleError::OperationError` `Display` format (#89)
When just running with the latest error changes from PR #80, I was
wondering why my error messages were disappearing but the `HRESULT` was
printed twice (this is on runs with `-WX` to treat warnings as errors).
Turns out I typo'd the `thiserror` format string to print the `HRESULT`
in the second place as well rather than the error blob, resulting in:
`Dxc error -80004005: -0x80004005`. Replacing that with `{1}` gives the
expected warning/error text from shader compilation again.