utils: Print status and error message when there's also a blob (#94)
What was initially added as a sanity assertion actually fails out in the
wild now: we have cases where DXIL shader validation fails, an error
code is returned, but there is also a blob with DXIL output (that may
not be valid per the error).

At least print this info as part of the assert, and maybe we should find
a better way to handle this "probably invalid" DXIL blob?

Perhaps just remove the assert - but also don't give the blob to the
caller if `get_status()` wasn't successful.