os: Fix "double negation" in `HRESULT` hexadecimal formatting (#91)
We typically see errors formatted as `-0x80004005`, because we tell
`pad_integral()` that the number is negative while we format-printed it
as **unsigned** integer which cannot be negative as represented by the
highest `0x80000000` bit being set - appearing as a double negation of
sorts? Either we remove the misleading `-` or print the number as
integer with the sign bit removed (which makes it harder to look up
online).

The end result is that we can and should just remove this strange
override of `LowerHex`, and rely purely on our `Display` implementation
of `HRESULT` that format-prints the inner `i32` as hexadecimal with `0x`
prefix, which automatically prints it as unsigned.

The end result is that we found one more case where `{:#x}` was
overridden to just `{:x}` because `LowerHex` is no longer implemented
for `HRESULT`, which was accidentally dropping the `0x` prefix in the
format string too.